# backend/server.js
/**
 * @file server.js
 * @description Express сервер для ранжування резюме.
 * @author Aliiev Omar
 */
const swaggerUi = require("swagger-ui-express");
const swaggerDocument = require("./swagger.json");
const express = require("express");
const cors = require("cors");
const bodyParser = require("body-parser");
const { TfIdf } = require("natural");

const app = express();
const PORT = 8000;

app.use(cors());
app.use(bodyParser.json());

/**
 * Список стоп-слів українською мовою.
 * Використовується для очищення тексту перед обробкою.
 */
const UKRAINIAN_STOP_WORDS = [
    "і", "й", "або", "але", "та", "що", "як", "в", "на", "з", "по", "за", "у", "із", "від", "до",
    "не", "так", "також", "чи", "ви", "ми", "вони", "це", "цей", "ця", "те", "ті", "бо", "хоча", "де", "коли"
];

/**
 * Очищує текст, видаляючи стоп-слова і переводячи в нижній регістр.
 * @param {string} text - Вхідний текст.
 * @returns {string} - Очищений текст.
 */
function cleanText(text) {
    return text
        .toLowerCase()
        .split(/\s+/)
        .filter(word => !UKRAINIAN_STOP_WORDS.includes(word))
        .join(" ");
}

/**
 * Обчислює косинусну схожість між двома векторами.
 * @param {number[]} vecA - Перший вектор.
 * @param {number[]} vecB - Другий вектор.
 * @returns {number} - Значення косинусної схожості (від 0 до 1).
 */
function cosineSimilarity(vecA, vecB) {
    const dotProduct = vecA.reduce((sum, a, i) => sum + a * vecB[i], 0);
    const magnitudeA = Math.sqrt(vecA.reduce((sum, a) => sum + a * a, 0));
    const magnitudeB = Math.sqrt(vecB.reduce((sum, b) => sum + b * b, 0));
    return magnitudeA && magnitudeB ? dotProduct / (magnitudeA * magnitudeB) : 0;
}

/**
 * Ранжує резюме на основі схожості з вакансією.
 * @param {Object} vacancy - Об'єкт вакансії.
 * @param {Object[]} resumes - Масив об'єктів резюме.
 * @returns {Object[]} - Масив ранжованих резюме з полями relevance_score і relevance_percentage.
 */
function rankResumes(vacancy, resumes) {
    const vacancyText = cleanText(
        `${vacancy.title} ${vacancy.description} ${vacancy.employment} ${vacancy.salary} ${vacancy.location}`
    );

    const resumeTexts = resumes.map(resume =>
        cleanText(`${resume.education} ${resume.work_experience} ${resume.skills} ${resume.personal_qualities} ${resume.about_me}`)
    );

    const tfidf = new TfIdf();
    tfidf.addDocument(vacancyText);
    resumeTexts.forEach(text => tfidf.addDocument(text));

    // Отримаємо всі унікальні терміни для створення однакових векторів
    const terms = tfidf.listTerms(0).map(term => term.term);

    const vacancyVector = terms.map(term => tfidf.tfidf(term, 0));

    const similarityScores = resumeTexts.map((_, i) => {
        const resumeVector = terms.map(term => tfidf.tfidf(term, i + 1));
        return cosineSimilarity(vacancyVector, resumeVector);
    });

    const maxRelevance = Math.max(...similarityScores);
    const maxPercentage = maxRelevance < 0.1 ? 70 : 100;

    const rankedResumes = resumes.map((resume, i) => ({
        id: resume.id,
        relevance_score: similarityScores[i],
        relevance_percentage: Math.round((similarityScores[i] / maxRelevance) * maxPercentage)
    }));

    return rankedResumes.sort((a, b) => b.relevance_score - a.relevance_score);
}


/**
 * Ранжує резюме за релевантністю до вакансії на основі TF-IDF та косинусної схожості.
 *
 * @name POST /rank_resumes_ai
 * @function
 * @param {Object} req.body - Об'єкт з вакансією та списком резюме.
 * @param {Object} req.body.vacancy - Об'єкт вакансії.
 * @param {string} req.body.vacancy.title - Назва вакансії.
 * @param {string} req.body.vacancy.description - Опис вакансії.
 * @param {string} req.body.vacancy.employment - Тип зайнятості.
 * @param {string} req.body.vacancy.salary - Зарплата.
 * @param {string} req.body.vacancy.location - Локація.
 * @param {Object[]} req.body.resumes - Масив резюме.
 * @param {number} req.body.resumes[].id - Ідентифікатор резюме.
 * @param {string} req.body.resumes[].education - Освіта кандидата.
 * @param {string} req.body.resumes[].work_experience - Досвід роботи.
 * @param {string} req.body.resumes[].skills - Навички.
 * @param {string} req.body.resumes[].personal_qualities - Особисті якості.
 * @param {string} req.body.resumes[].about_me - Про себе.
 * @param {string} req.body.resumes[].date - Дата подання резюме (у форматі ISO).
 * @returns {Object[]} Масив ранжованих резюме з оцінками релевантності.
 * @returns {number} return[].id - Ідентифікатор резюме.
 * @returns {number} return[].relevance_score - Значення косинусної схожості.
 * @returns {number} return[].relevance_percentage - Відсоток релевантності від 0 до 100.
 *
 * @example
 * POST /rank_resumes_ai
 * Content-Type: application/json
 * {
 *   "vacancy": { "title": "Java Developer", "description": "...", "employment": "full-time", "salary": "1000", "location": "Kyiv" },
 *   "resumes": [
 *     { "id": 1, "education": "...", "work_experience": "...", "skills": "...", "personal_qualities": "...", "about_me": "...", "date": "2023-05-01" }
 *   ]
 * }
 */
app.post("/rank_resumes_ai/", (req, res) => {
    try {
        const { vacancy, resumes } = req.body;
        const ranked = rankResumes(vacancy, resumes);
        res.json(ranked);
    } catch (err) {
        console.error("Помилка AI ранжування:", err);
        res.status(500).json({ error: err.toString() });
    }
});

/**
 * Сортує список резюме за датою подання (від найновіших до найстаріших).
 *
 * @name POST /rank_resumes_date
 * @function
 * @param {Object} req.body - Об'єкт з масивом резюме.
 * @param {Object[]} req.body.resumes - Масив резюме.
 * @param {number} req.body.resumes[].id - Ідентифікатор резюме.
 * @param {string} req.body.resumes[].education - Освіта кандидата.
 * @param {string} req.body.resumes[].work_experience - Досвід роботи.
 * @param {string} req.body.resumes[].skills - Навички.
 * @param {string} req.body.resumes[].personal_qualities - Особисті якості.
 * @param {string} req.body.resumes[].about_me - Про себе.
 * @param {string} req.body.resumes[].date - Дата подання резюме (у форматі ISO).
 * @returns {Object[]} Відсортований масив резюме за датою.
 *
 * @example
 * POST /rank_resumes_date
 * Content-Type: application/json
 * {
 *   "resumes": [
 *     { "id": 1, "education": "...", "work_experience": "...", "skills": "...", "personal_qualities": "...", "about_me": "...", "date": "2023-05-01" }
 *   ]
 * }
 */
app.post("/rank_resumes_date/", (req, res) => {
    try {
        const { resumes } = req.body;
        const sorted = resumes.sort((a, b) => new Date(b.date) - new Date(a.date));
        res.json(sorted);
    } catch (err) {
        console.error("Помилка сортування по даті:", err);
        res.status(500).json({ error: err.toString() });
    }
});

/**
 * Запускає сервер на заданому порту.
 */

app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
app.listen(PORT, () => {
    console.log(`Server running at http://localhost:${PORT}`);
});